회원수정 불가 오류

+++ ClubManagement.Common/Hlepers/DataConvertHelper.cs
@@ -4,19 +4,19 @@ namespace ClubManagement.Common.Hlepers
 {
     public static class DataConvertHelper
     {
-        public static decimal ConvertRate(object denominator, object numerator)
+        public static double ConvertRate(object denominator, object numerator)
         {
-            if (!decimal.TryParse(Convert.ToString(denominator), out decimal _denominator))
+            if (!double.TryParse(Convert.ToString(denominator), out double _denominator))
                 _denominator = 0;
 
-            if (!decimal.TryParse(Convert.ToString(numerator), out decimal _numerator))
+            if (!double.TryParse(Convert.ToString(numerator), out double _numerator))
                 _numerator = 0;
 
             if (_denominator == 0)
                 return _numerator > 0 ? 100 : 0;
 
 
-            return Math.Round((_numerator / _denominator) * 100, 2);
+            return Math.Round((double)(_numerator / _denominator) * 100, 2);
         }
     }
 }
